requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14,<3.15
requests>=2.31,<3.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0 